added authentication

+++ app/Http/Controllers/PostsController.php
@@ -5,9 +5,21 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Post;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Auth;
 
 class PostsController extends Controller
 {
+
+    /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth', ['except'=>['index', 'show']]);
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -52,6 +64,7 @@ class PostsController extends Controller
         $post = new Post();
         $post->title = $request->input('title');
         $post->body = $request->input('body');
+        $post->user_id = auth()->user()->id;
         $post->save();
 
         return redirect('/posts')->with('success', 'Post created');
@@ -79,6 +92,12 @@ class PostsController extends Controller
     {
         //
         $post = Post::find($id);
+
+        // check for correct user
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('/posts')->withErrors('Unauthorized');
+        }
+
         return view('posts.edit')->withPost($post);
     }
 
@@ -115,6 +134,11 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+        // check for correct user
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('/posts')->withErrors('Unauthorized');
+        }
+
         $post->delete();
 
         return redirect('/posts')->withSuccess('Post successfully deleted');
